+++ hungarian.hpp
@@ -32,7 +32,7 @@ public:
     double hungarian_solve();
 
     void adjust_matrix_by_min_uncovered_num(std::vector<std::vector<double>>& matrix,
-            std::vector<int> covered_rows, std::vector<int> covered_columns);
+            std::vector<size_t> covered_rows, std::vector<size_t> covered_columns);
 
     std::pair<std::vector<int>, std::vector<int>> find_matches(const std::vector<std::vector<bool>>& zero_locations);
     static std::pair<std::vector<int>, std::vector<int>> mark_rows_and_columns(
@@ -62,12 +62,12 @@ private:
     std::vector<std::vector<bool>> zero_locations;
 
     size_t square_matrix_size;
-    std::vector<int> marked_rows;
-    std::vector<int> marked_columns;
+    std::vector<size_t> marked_rows;
+    std::vector<size_t> marked_columns;
 
     std::vector<std::vector<bool>> choices;
-    std::vector<int> covered_rows;
-    std::vector<int> covered_columns;
+    std::vector<size_t> covered_rows;
+    std::vector<size_t> covered_columns;
 
     double epsilon;
 
@@ -88,9 +88,9 @@ public:
 
     int find_marked_column_without_choice();
 
-    std::vector<int> get_covered_rows();
+    std::vector<size_t> get_covered_rows();
 
-    std::vector<int> get_covered_columns();
+    std::vector<size_t> get_covered_columns();
 
     // debug func
     bool is_double_equal(double x, double y) const
